+++ .gitignore
@@ -10,6 +10,7 @@ bin/
 # Gradle
 .gradle
 build
+mappings/
 
 # Visual Studio code
 .vscode*
