+++ src/main/java/org/researchspace/sail/rest/RESTSailConfig.java
@@ -42,7 +42,7 @@ import com.beust.jcommander.internal.Maps;
  *
  */
 
-public class RESTSailConfig extends AbstractRESTWrappingSailConfig {
+public class RESTSailConfig extends AbstractServiceWrappingSailConfig {
 
     private String httpMethod;
 
@@ -60,6 +60,13 @@ public class RESTSailConfig extends AbstractRESTWrappingSailConfig {
      * Additional HTTP headers
      */
     private Map<String, String> httpHeaders;
+
+    /**
+     * request per second
+     */
+    private Integer requestRateLimit;
+
+    private String userAgent;
 
     public RESTSailConfig() {
         super(RESTSailFactory.SAIL_TYPE);
@@ -90,35 +97,47 @@ public class RESTSailConfig extends AbstractRESTWrappingSailConfig {
             if (name.isPresent() && value.isPresent())
                 httpHeaders.put(name.get().stringValue(), value.get().stringValue());
         });
+
+        Models.objectLiteral(model.filter(implNode, MpRepositoryVocabulary.REQUEST_RATE_LIMIT, null))
+                .ifPresent(lit -> setRequestRateLimit(lit.intValue()));
+        Models.objectLiteral(model.filter(implNode, MpRepositoryVocabulary.USER_AGENT, null))
+                .ifPresent(lit -> setUserAgent(lit.stringValue()));
     }
 
     @Override
     public Resource export(Model model) {
         Resource implNode = super.export(model);
+        var vf = SimpleValueFactory.getInstance();
 
         // Store the HTTP method in the model
-        if (!StringUtils.isEmpty(getHttpMethod()))
-            model.add(implNode, MpRepositoryVocabulary.HTTP_METHOD,
-                    SimpleValueFactory.getInstance().createLiteral(getHttpMethod()));
+        if (!StringUtils.isEmpty(getHttpMethod())) {
+            model.add(implNode, MpRepositoryVocabulary.HTTP_METHOD, vf.createLiteral(getHttpMethod()));
+        }
 
-        if (!StringUtils.isEmpty(getInputFormat()))
-            model.add(implNode, MpRepositoryVocabulary.INPUT_FORMAT,
-                    SimpleValueFactory.getInstance().createLiteral(getInputFormat()));
+        if (!StringUtils.isEmpty(getInputFormat())) {
+            model.add(implNode, MpRepositoryVocabulary.INPUT_FORMAT, vf.createLiteral(getInputFormat()));
+        }
 
-        if (!StringUtils.isEmpty(getMediaType()))
-            model.add(implNode, MpRepositoryVocabulary.MEDIA_TYPE,
-                    SimpleValueFactory.getInstance().createLiteral(getMediaType()));
+        if (!StringUtils.isEmpty(getMediaType())) {
+            model.add(implNode, MpRepositoryVocabulary.MEDIA_TYPE, vf.createLiteral(getMediaType()));
+        }
 
         if (!Objects.isNull(httpHeaders) && httpHeaders.size() > 0) {
             for (Map.Entry<String, String> header : httpHeaders.entrySet()) {
-                BNode root = SimpleValueFactory.getInstance().createBNode();
+                BNode root = vf.createBNode();
 
                 model.add(implNode, MpRepositoryVocabulary.HTTP_HEADER, root);
-                model.add(root, MpRepositoryVocabulary.NAME,
-                        SimpleValueFactory.getInstance().createLiteral(header.getKey()));
-                model.add(root, MpRepositoryVocabulary.VALUE,
-                        SimpleValueFactory.getInstance().createLiteral(header.getValue()));
+                model.add(root, MpRepositoryVocabulary.NAME, vf.createLiteral(header.getKey()));
+                model.add(root, MpRepositoryVocabulary.VALUE, vf.createLiteral(header.getValue()));
             }
+        }
+
+        if (getRequestRateLimit() != null) {
+            model.add(implNode, MpRepositoryVocabulary.REQUEST_RATE_LIMIT, vf.createLiteral(getRequestRateLimit()));
+        }
+
+        if (getUserAgent() != null) {
+            model.add(implNode, MpRepositoryVocabulary.USER_AGENT, vf.createLiteral(getUserAgent()));
         }
 
         return implNode;
@@ -150,5 +169,21 @@ public class RESTSailConfig extends AbstractRESTWrappingSailConfig {
 
     public Map<String, String> getHttpHeaders() {
         return this.httpHeaders;
+    }
+
+    public Integer getRequestRateLimit() {
+        return requestRateLimit;
+    }
+
+    protected void setRequestRateLimit(int requestRateLimit) {
+        this.requestRateLimit = requestRateLimit;
+    }
+
+    public String getUserAgent() {
+        return userAgent;
+    }
+
+    protected void setUserAgent(String userAgent) {
+        this.userAgent = userAgent;
     }
 }
